+++ packages/eyes-sdk-core/lib/images/MutableImage.js
@@ -248,8 +248,10 @@ class MutableImage {
     const srcImageBmp = await srcImage.getImageData()
     let width = srcImage.getWidth()
     let height = srcImage.getHeight()
-    const maxWidth = this.getWidth() - dx
-    const maxHeight = this.getHeight() - dy
+    const x = Math.round(dx)
+    const y = Math.round(dy)
+    const maxWidth = this.getWidth() - x
+    const maxHeight = this.getHeight() - y
 
     if (maxWidth < width) {
       width = maxWidth
@@ -259,13 +261,7 @@ class MutableImage {
       height = maxHeight
     }
 
-    ImageUtils.copyPixels(
-      this._imageBmp,
-      {x: dx, y: dy},
-      srcImageBmp,
-      {x: 0, y: 0},
-      {width, height},
-    )
+    ImageUtils.copyPixels(this._imageBmp, {x, y}, srcImageBmp, {x: 0, y: 0}, {width, height})
   }
 
   /**