+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## Unreleased
 
+- fix mutable image round position
 
 ## 12.16.2 - 2021/3/16
 
-- fix setViewportSize
 - fix setViewportSize
 ## 12.16.1 - 2021/3/15